refactor: rename repository package to infra
