+++ src/main/java/com/example/minimart/order/infra/OrderItemJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.minimart.order.repository;
+package com.example.minimart.order.infra;
 
-import com.example.minimart.order.repository.entity.OrderItem;
+import com.example.minimart.order.infra.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderItemJpaRepository extends JpaRepository<OrderItem, Long> {
