+++ src/main/java/com/example/minimart/order/infra/OrderJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.minimart.order.repository;
+package com.example.minimart.order.infra;
 
-import com.example.minimart.order.repository.entity.Order;
+import com.example.minimart.order.infra.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderJpaRepository extends JpaRepository<Order, Long> {
