+++ src/main/java/com/example/minimart/order/service/OrderService.java
@@ -1,11 +1,11 @@
 package com.example.minimart.order.service;
 
 import com.example.minimart.order.controller.dto.request.CreateOrderRequest;
-import com.example.minimart.order.repository.OrderItemJpaRepository;
-import com.example.minimart.order.repository.OrderJpaRepository;
-import com.example.minimart.order.repository.entity.Order;
-import com.example.minimart.order.repository.entity.OrderItem;
-import com.example.minimart.order.repository.entity.OrderStatus;
+import com.example.minimart.order.infra.OrderItemJpaRepository;
+import com.example.minimart.order.infra.OrderJpaRepository;
+import com.example.minimart.order.infra.entity.Order;
+import com.example.minimart.order.infra.entity.OrderItem;
+import com.example.minimart.order.infra.entity.OrderStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
